Fix bug, add new instruction
set_z80_pc_abs was writing the high byte of the logical PC to
z80_pc_logical not z80_pc_logical + 1. I've reworking this so the
logical PC only has a high byte, in the hope it will be clearer. The low
byte is shared with the physical PC and lives in Y.